refactor(client): extract HoverActionButton from message_item (6/7)
Per .claude/plans/generic-growing-bentley.md. Continues the message/
folder pattern.

- New: widgets/message/hover_action_button.dart (56 LoC) — the 28×28
  visual chip with 44×44 a11y hit target used by the hover-action bar
  on every message. Now reusable from the long-press action sheet on
  mobile and any other surface that wants the same look.
- message_item.dart: removes the file-internal _HoverActionButton class
  (~45 LoC) and renames the two call sites.

Net: 2058 -> 2014 LoC in the parent. flutter analyze clean.
flutter test test/widgets/message_item_test.dart — all 37 pass.
